🎨 chore(client): update package.json with new dependencies The following dependencies were added to the client's package.json file: chart.js, chartjs-adapter-date-fns, and react-chartjs-2. These dependencies are required for the new DynamicBarChart component that was added to the Activity view.
🚀 feat(client): add DynamicBarChart component to Activity view
The DynamicBarChart component was added to the Activity view, which displays a bar chart of sales per month. The chart is populated with dummy data for now. The component uses the chart.js library and its adapter for date-fns. The chart is displayed above the DynamicCarousel component.

🗑️ chore(client): remove unused SalesActivity component
The SalesActivity component was removed from the client's codebase as it was no longer being used.

🚚 chore(client): move Scheduling component to views folder
The Scheduling component was moved from the components folder to the views folder

✨ feat(Scheduling.jsx): add Scheduling component
The Scheduling component is added to the application. It contains a form to add events, a calendar to display events, and a timeline to show scheduled events. The component uses React Bootstrap for styling and moment.js for date formatting. The component uses state to manage events and passes the events to the calendar and timeline components.