🚀 chore(client): update dependencies in package.json The dependencies "@mui/material" and "mongoose" were updated to their latest versions. Additionally, the "react-photo-gallery" package was added as a new dependency.
🎨 style(client): add PhotoTile.module.css
A new CSS module file was added to the client/src/Style.modules.css directory. It contains styles for the PhotoTile component.

🎨 style(client): remove PhotoTile.jsx
The PhotoTile component was removed from the project, and its corresponding file was deleted.

🎨 style(client): add Figure.jsx
A new component, Figure, was added to the project. It renders an image with a fixed width and height.

🎨 style(client): add Gallery.jsx
A new component, Gallery, was added to the project. It renders a grid of images.

🎨 style(client): update Table.jsx
The DynamicTable component was updated to pass a handleDetailsClick function

🔨 refactor(LeadDetails.jsx, Leads.jsx, Listings.jsx): rename Copy component to CopyButton and add handleDetailsClick function
The Copy component has been renamed to CopyButton to better reflect its purpose. Additionally, a handleDetailsClick function has been added to the Leads and Listings components to handle the click event when a user wants to view the details of a lead or listing. This function logs the ID of the selected item and navigates to the corresponding details page.

🎉 feat(ListingsDetails.jsx): add ListingDetails component
✨ feat(Main.jsx): add Listings and ListingDetails routes
The ListingDetails component is added to display the details of a house listing. The component includes a modal that displays a carousel of images. The Main component now includes the Listings and ListingDetails routes. The Listings route displays a table of house listings, and the ListingDetails route displays the details of a specific house listing.