🚀 chore(client): add Mantine UI library Mantine UI library has been added to the client's package.json file. This library will be used to improve the user interface of the application.
🐛 fix(client): remove console.log statement
A console.log statement has been removed from the DynamicTable component's code. This statement was used for debugging purposes and is no longer needed.

🐛 fix(client): fix formatting issue in Leads.jsx
A formatting issue has been fixed in the validations array in the Leads.jsx file. The array was not properly formatted, causing an error.

🐛 fix(client): fix formatting issue in Main.jsx
A formatting issue has been fixed in the houseListings array in the Main.jsx file. The array was not properly formatted, causing an error.

🚀 chore(client): add gray.todo file
A gray.todo file has been added to the client's views directory. This file will be used to keep track of