🎨 style(Table.module.css): remove unused CSS classes 🚀 feat(Calculator.jsx): add RealEstateCalculator component The unused CSS classes were removed to improve code readability and maintainability. The RealEstateCalculator component was added to the project, which allows users to calculate the net income, cash on cash return, and cap rate of a real estate investment property. The component uses React Bootstrap Form and Button components to take user input and display the results of the calculations.
🎨 style(SideNav.jsx): add hover effect to buttons and change button styles
🎨 style(Table.jsx): add hover effect to buttons and change button styles
The buttons in the SideNav and Table components now have a hover effect that changes the background color and text color. The button styles have also been updated to improve consistency and readability.

🎨 style(Activity.jsx): add a calculator component and adjust layout
The layout of the Activity component has been adjusted to include a new calculator component. The calculator component has been added to the left side of the DynamicBarChart component. The layout has been adjusted using the react-bootstrap Col and Row components.