🐛 fix(Leads.jsx): fix state variable name for all leads of logged user ✨ feat(Leads.jsx): add axios import and function to get all leads for logged user The state variable for all leads of the logged user was incorrectly named, causing the data to not be displayed properly. The variable was renamed to `allLeadsforLoggedUser`. Additionally, an axios import was added and a new function was created to query the server for all leads of the currently logged in user. This allows the application to display all leads for the logged in user, rather than just the leads passed in as a prop.
🐛 fix(Login.jsx): add state variables and axios post request to handle form submission
✨ feat(Login.jsx): add error handling for login form submission
State variables were added to handle the email and password form inputs, as well as to store any errors that may occur during form submission. An axios post request was added

🐛 fix(user.controller.js): fix error object in response
✨ feat(user.controller.js): add loggedUser object to login response
✨ feat(user.routes.js): add checkUserSession route
The error object in the response of the tryLoginOneUserByEmail function was not being properly handled. The error object was being spread into the response object, which caused the message property to be overwritten. The error object is now being properly handled. The login response now includes a loggedUser object, which is a copy of the existingUser object with the _id property converted to a string. The checkUserSession route was added to allow the front-end to check if a user is logged in.